+++ CHANGES.md
@@ -48,6 +48,20 @@ To be released.
 [#195]: https://github.com/dahlia/fedify/issues/195
 
 
+Version 1.3.5
+-------------
+
+Released on January 21, 2025.
+
+ -  Fixed a bug where `CreateFederationOptions.allowPrivateAddress` option had
+    been ignored by the `Context.lookupObject()` method when it had taken
+    a fediverse handle.
+
+ -  The `lookupWebFinger()` function became to silently return `null` when
+    it fails to fetch the WebFinger document due to accessing a private network
+    address, instead of throwing a `UrlError`.
+
+
 Version 1.3.4
 -------------